requirements.txt
streamlit>=1.27.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=1.10.0,<2.0.0
langchain==0.0.335
langchain-community==0.0.13
langchain-openai==0.0.5
duckduckgo-search>=3.9.0
requests>=2.28.0
beautifulsoup4>=4.12.0
langgraph>=0.0.11
setuptools>=65.5.1
openai>=1.1.1
httpx==0.23.3
aiohttp>=3.8.0
# Ollama support
llama-cpp-python>=0.2.0